Fix git comments
